Made the following changes to Spores:
* wrote a custom Transformer to correctly transform path-dependent types
that depend on captured variables and function parameters. The changes
are: method transformTypes in MacroImpl, calls to the method in check,
check2 and checkNullary. The rationale for these changes are the tests
pTTParameterCapture, pTTCapturedCapture in run.basic.Basic.

* changed the spore structure so that captured values are 'val'
rather than 'var. With these changes, path-dependent types can be formed
with the captured values. Test pTTCapturedCapture shows the need for this.
The 'captured' value is now provided as a constructor argument.
The code generated by check, check2, checkNullary has changed.
spores-pickling had to be changed to set 'captured' by reflection to
avoid assignment to val.